requirements.txt
streamlit>=1.30.0
yfinance>=0.2.40
pandas>=2.0.0
plotly>=5.18.0
gspread>=5.12.0
google-auth>=2.20.0
